@@ -30,9 +30,10 @@
     "lodash": "^4.17.4",
     "node-sass": "^4.5.0",
     "postcss-loader": "^1.2.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-test-renderer": "^15.4.2",
+    "prop-types": "^15.5.0",
+    "react": "^15.5.0",
+    "react-dom": "^15.0.0",
+    "react-test-renderer": "^15.0.0",
     "sass-loader": "^4.1.1",
     "style-loader": "^0.13.1",
     "webpack": "^2.2.1",
